tools/pipeline_runner/pyproject.toml
[tool.poetry]
name = "pipeline-runner"
version = "0.1.0"
description = "CI/CD and operational pipelines for the Openclaw Librarian Agent"
authors = ["Redlex Gilgamesh"]
packages = [{include = "pipeline_runner"}]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8.1"
pyyaml = "^6.0"
rich = "^13.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0"
pytest-cov = "^6.0"
ruff = "^0.8"

[tool.poetry.scripts]
pipeline = "pipeline_runner.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short --cov=pipeline_runner --cov-report=term-missing"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
